pyproject.toml
[tool.poetry]
name = "ecommerce-back"
version = "0.1.0"
description = ""
authors = ["Jordana Loreto <jordanaloreto45@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
